+++ ubo_app/constants.py
@@ -36,7 +36,7 @@ GRPC_ENVOY_LISTEN_ADDRESS = os.environ.get('UBO_GRPC_ENVOY_LISTEN_ADDRESS', '127
 GRPC_ENVOY_LISTEN_PORT = int(os.environ.get('UBO_GRPC_ENVOY_LISTEN_PORT', '50052'))
 
 WEB_UI_LISTEN_ADDRESS = os.environ.get('UBO_WEB_UI_LISTEN_ADDRESS', '0.0.0.0')  # noqa: S104
-WEB_UI_LISTEN_PORT = int(os.environ.get('UBO_WEB_UI_LISTEN_PORT', '21215'))
+WEB_UI_LISTEN_PORT = int(os.environ.get('UBO_WEB_UI_LISTEN_PORT', '4321'))
 WEB_UI_DEBUG_MODE = str_to_bool(os.environ.get('UBO_WEB_UI_DEBUG_MODE', 'False')) == 1
 
 UPDATE_ASSETS_PATH = Path(f'{INSTALLATION_PATH}/_update/')
@@ -58,7 +58,6 @@ PICOVOICE_ACCESS_KEY = 'PICOVOICE_ACCESS_KEY'
 
 DEBUG_MODE_DOCKER = str_to_bool(os.environ.get('UBO_DEBUG_DOCKER', 'False')) == 1
 DOCKER_CREDENTIALS_TEMPLATE = 'DOCKER_CREDENTIALS_{}'
-DOCKER_PREFIX = os.environ.get('UBO_DOCKER_PREFIX', '')
 
 CONFIG_PATH = platformdirs.user_config_path(appname='ubo', ensure_exists=True)
 SECRETS_PATH = CONFIG_PATH / '.secrets.env'