+++ ubo_app/utils/log_process.py
@@ -11,25 +11,32 @@ from ubo_app.store.services.notifications import (
 )
 
 
-async def log_async_process(process: Process) -> None:
+async def log_async_process(
+    process: Process,
+    *,
+    title: str = 'Error',
+    message: str,
+) -> None:
     """Log the output of the process if it fails.
 
     Arguments:
     ---------
     process: Process - The process to log the output of.
+    title: str - The title of the notification.
+    message: str - The message to display in the notification.
 
     """
     if process.returncode != 0:
         logs = ''
         if process.stdout:
-            logs += 'stdout:\n' + (await process.stdout.read()).decode()
+            logs += '---\nstdout:\n' + (await process.stdout.read()).decode()
         if process.stderr:
-            logs += 'stderr:\n' + (await process.stderr.read()).decode()
+            logs += '---\nstderr:\n' + (await process.stderr.read()).decode()
         store.dispatch(
             NotificationsAddAction(
                 notification=Notification(
-                    title='Docker Composition Error',
-                    content='Failed to run the composition',
+                    title=title,
+                    content=message,
                     extra_information=NotificationExtraInformation(
                         text=logs,
                     ),